+++ packages/_infra/src/serve/lambda.tiler.ts
@@ -37,6 +37,8 @@ export class LambdaTiler extends Construct {
       code: lambda.Code.fromAsset(CODE_PATH),
       environment: {
         [Env.PublicUrlBase]: config.PublicUrlBase,
+        [Env.AwsRoleConfigBucket]: config.AwsRoleConfigBucket,
+        [Env.ImportImageryBucket]: config.ImportImageryBucket,
         AWS_NODEJS_CONNECTION_REUSE_ENABLED: '1',
       },
       logRetention: RetentionDays.ONE_MONTH,
@@ -45,5 +47,7 @@ export class LambdaTiler extends Construct {
       const cogBucket = s3.Bucket.fromBucketName(this, `CogBucket${bucketName}`, bucketName);
       cogBucket.grantRead(this.lambda);
     }
+    const configBucket = s3.Bucket.fromBucketName(this, 'ConfigBucket', config.AwsRoleConfigBucket);
+    configBucket.grantRead(this.lambda);
   }
 }