+++ packages/cli/src/cli/cogify/__test__/action.batch.test.ts
@@ -1,6 +1,6 @@
 import o from 'ospec';
 import { CogJob } from '../../../cog/types.js';
-import { ActionBatchJob, extractResolutionFromName } from '../action.batch.js';
+import { BatchJob, extractResolutionFromName } from '../batch.job.js';
 
 o.spec('action.batch', () => {
   o('extractResolutionFromName', () => {
@@ -15,17 +15,17 @@ o.spec('action.batch', () => {
 
   o('should create valid jobNames', () => {
     const fakeJob = { id: '01FHRPYJ5FV1XAARZAC4T4K6MC', name: 'geographx_nz_texture_shade_2012_8-0m' } as CogJob;
-    o(ActionBatchJob.id(fakeJob, '0')).equals('01FHRPYJ5FV1XAARZAC4T4K6MC-9af5e139bbb3e502-0');
+    o(BatchJob.id(fakeJob, '0')).equals('01FHRPYJ5FV1XAARZAC4T4K6MC-9af5e139bbb3e502-0');
 
     fakeJob.name = 'ōtorohanga_urban_2021_0.1m_RGB';
-    o(ActionBatchJob.id(fakeJob, '0')).equals('01FHRPYJ5FV1XAARZAC4T4K6MC-5294acface81c107-0');
+    o(BatchJob.id(fakeJob, '0')).equals('01FHRPYJ5FV1XAARZAC4T4K6MC-5294acface81c107-0');
   });
 
   o('should truncate job names to 128 characters', () => {
     const fakeJob = { id: '01FHRPYJ5FV1XAARZAC4T4K6MC', name: 'geographx_nz_texture_shade_2012_8-0m' } as CogJob;
 
     o(
-      ActionBatchJob.id(
+      BatchJob.id(
         fakeJob,
         'this is a really long file name it should over flow 128 characters so it should be truncated at some point.tiff',
       ),