+++ packages/cli/src/cog/cog.stac.job.ts
@@ -76,6 +76,11 @@ export interface JobCreationContext {
      * @Default  GdalCogBuilderDefaults.resampling
      */
     resampling?: GdalCogBuilderResampling;
+
+    /**
+     * Override job processing Id
+     */
+    processingId?: string;
   };
 
   /**
@@ -112,6 +117,7 @@ export class CogStacJob implements CogJob {
     gsd: number;
     zoom: number;
   };
+  processingId: string;
 
   /**
      * Load the job.json
@@ -211,6 +217,9 @@ export class CogStacJob implements CogJob {
         oneCogCovering: ctx.oneCogCovering,
       },
     });
+
+    // Only include processing ID if exists
+    if (ctx.override?.processingId != null) job.processingId = ctx.override?.processingId;
 
     const nowStr = new Date().toISOString();