+++ packages/cli/src/cog/types.ts
@@ -68,6 +68,8 @@ export interface CogJobJson {
 
   source: CogSourceProperties;
   output: CogOutputProperties;
+
+  processingId?: string;
 }
 
 export interface CogJob extends CogJobJson {