+++ packages/lambda-tiler/src/routes/import.ts
@@ -1,11 +1,12 @@
 import { HttpHeader, LambdaHttpRequest, LambdaHttpResponse } from '@linzjs/lambda';
-import { Config, Const, fsa } from '@basemaps/shared';
+import { Config, fsa } from '@basemaps/shared';
 import { createHash } from 'crypto';
 import { findImagery, RoleRegister } from '../import/imagery.find.js';
 import { Nztm2000Tms, TileMatrixSets } from '@basemaps/geo';
 import { getJobCreationContext } from '../import/make.cog.js';
-import { ConfigProcessingJob, ConfigProviderDynamo } from '@basemaps/config';
+import { ConfigProcessingJob, JobStatus } from '@basemaps/config';
 import { CogJobFactory } from '@basemaps/cli';
+import * as ulid from 'ulid';
 
 /**
  * Trigger import imagery job by this endpoint
@@ -27,33 +28,39 @@ export async function Import(req: LambdaHttpRequest): Promise<LambdaHttpResponse
   }
 
   // Find the imagery from s3
-  if (path == null || !path.startsWith('s3://')) return new LambdaHttpResponse(500, 'Invalid s3 path');
+  if (path == null || !path.startsWith('s3://')) return new LambdaHttpResponse(400, `Invalid s3 path: ${path}`);
   const role = await RoleRegister.findRole(path);
-  if (role == null) return new LambdaHttpResponse(500, 'Unable to Access the bucket');
+  if (role == null) return new LambdaHttpResponse(403, 'Unable to Access the s3 bucket');
   const files = await findImagery(path);
   if (files.length === 0) return new LambdaHttpResponse(404, 'Imagery Not Found');
 
   // Prepare Cog jobs
   const ctx = await getJobCreationContext(path, targetTms, imageryName, role, files);
 
-  const id = createHash('sha256').update(JSON.stringify(ctx)).digest('base64');
-  const jobId = Config.ProcessingJob.id(id);
+  const hash = createHash('sha256').update(JSON.stringify(ctx)).digest('base64');
+  const jobId = Config.ProcessingJob.id(hash);
   let jobConfig = await Config.ProcessingJob.get(jobId);
   if (jobConfig == null) {
-    // Add id back to JobCreationContext
+    // Add ids into JobCreationContext
+    const id = ulid.ulid();
     ctx.override!.id = id;
+    ctx.override!.processingId = jobId;
     ctx.outputLocation.path = fsa.join(ctx.outputLocation.path, id);
 
-    // Start processing job
-    await CogJobFactory.create(ctx);
+    // Insert Processing job config
     jobConfig = {
       id: jobId,
       name: path,
-      status: 'processing',
+      status: JobStatus.Processing,
+      tileMatrix: targetTms.identifier,
+      tileSet: Config.TileSet.id(id),
     } as ConfigProcessingJob;
 
-    const config = new ConfigProviderDynamo(Const.TileMetadata.TableName);
-    await config.ProcessingJob.put(jobConfig);
+    if (Config.ProcessingJob.isWriteable()) await Config.ProcessingJob.put(jobConfig);
+    else return new LambdaHttpResponse(403, 'Unable to insert the Processing Job config');
+
+    // Start processing job
+    await CogJobFactory.create(ctx);
   }
 
   const json = JSON.stringify(jobConfig);