+++ packages/landing/src/components/debug.tsx
@@ -33,12 +33,16 @@ export class Debug extends Component<
   }
 
   waitForMap = (): void => {
-    if (this.props.map == null) {
+    const map = this.props.map;
+    if (map == null) {
       setTimeout(this.waitForMap, 20);
       return;
     }
-    this.props.map.resize();
-    this.props.map.once('load', () => {
+
+    (window as any).MaplibreMap = map;
+
+    map.resize();
+    map.once('load', () => {
       Config.map.on('change', () => {
         if (this.props.map == null) return;
         const locationHash = WindowUrl.toHash(Config.map.getLocation(this.props.map));
@@ -47,6 +51,13 @@ export class Debug extends Component<
         this.updateFromConfig();
       });
       this.updateFromConfig();
+
+      // Jam a div into the page once the map has loaded so tools like playwright can see the map has finished loading
+      if (Config.map.debug['debug.screenshot']) {
+        const loadedDiv = document.createElement('div');
+        loadedDiv.id = 'map-loaded';
+        document.body.appendChild(loadedDiv);
+      }
     });
   };
 
@@ -74,15 +85,22 @@ export class Debug extends Component<
           <div className="debug__value">{Config.map.tileMatrix.identifier}</div>
         </div>
         {this.renderSliders()}
-        <div className="debug__info">
-          <label className="debug__label">Purple</label>
-          <input
-            type="checkbox"
-            onClick={this.togglePurple}
-            checked={Config.map.debug['debug.background'] === 'magenta'}
-          />
-        </div>
+        {this.renderPurple()}
         {this.renderSourceToggle()}
+      </div>
+    );
+  }
+
+  renderPurple(): ComponentChild | null {
+    if (Config.map.debug['debug.screenshot']) return;
+    return (
+      <div className="debug__info">
+        <label className="debug__label">Purple</label>
+        <input
+          type="checkbox"
+          onClick={this.togglePurple}
+          checked={Config.map.debug['debug.background'] === 'magenta'}
+        />
       </div>
     );
   }
@@ -216,6 +234,8 @@ export class Debug extends Component<
   }
 
   renderSliders(): ComponentChild | null {
+    // Disable the sliders for screenshots
+    if (Config.map.debug['debug.screenshot']) return;
     // Only 3857 currently works with OSM/Topographic map
     if (Config.map.tileMatrix.identifier !== GoogleTms.identifier) {
       return (
@@ -277,6 +297,8 @@ export class Debug extends Component<
 
     const layers = styleJson.layers?.filter((f) => f.type !== 'custom' && f.source === 'LINZ Basemaps') ?? [];
 
+    // Do not hide topographic layers when trying to inspect the topographic layer
+    if (Config.map.layerId === 'topographic') return;
     // Force all the layers to be invisible to start, otherwise the map will "flash" on then off
     for (const layer of layers) {
       if (layer.type === 'custom') continue;
@@ -323,7 +345,6 @@ export class Debug extends Component<
   };
 
   adjustRaster(rasterId: 'osm' | 'linz-aerial', range: number): void {
-    // Config.map.setDebug(`debug-layer-${rasterId}`, range);
     if (this.props.map.getSource(rasterId) == null) {
       this.props.map.addSource(rasterId, {
         type: 'raster',
@@ -347,6 +368,13 @@ export class Debug extends Component<
         maxzoom: 24,
         paint: { 'raster-opacity': 0 },
       });
+
+      // Ensure this raster layers are below the vector layer
+      const sourceLayerId = `${Config.map.layerId}_source_fill`;
+      const isSourceLayerEnabled = this.props.map.getLayer(sourceLayerId) != null;
+      if (isSourceLayerEnabled) {
+        this.props.map.moveLayer(rasterId, sourceLayerId);
+      }
     }
     this.props.map.setPaintProperty(rasterId, 'raster-opacity', range);
   }